% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDatasets.R
\name{addDatasets}
\alias{addDatasets}
\title{addDatasets}
\usage{
addDatasets(miceObj, datasets = 3, parallel = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{miceObj}{A miceDefs object created by \code{miceRanger}.}

\item{datasets}{The number of iterations to run the MICE algorithm on each dataset.}

\item{parallel}{Should the process run in parallel? This process will take advantage of any cluster 
set up when \code{miceRanger} is called.}

\item{verbose}{should progress be printed?}

\item{...}{other parameters passed to \code{ranger()} to control model building.}
}
\value{
a miceDefs object with additional iterations.
}
\description{
Add datasets to a current miceDefs object. 
Adds the same number of iterations as other datasets.
}
\examples{
data(iris)

ampIris <- amputeData(iris)

miceObj <- miceRanger(
    ampIris
  , m = 2
  , maxiter = 2
  , verbose=FALSE
)
miceObj <- addIterations(
    miceObj
  , iters=2
  , verbose=FALSE
)
}
