% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amputeData.R
\name{amputeData}
\alias{amputeData}
\title{amputeData}
\usage{
amputeData(data, perc = 0.1, cols = names(data))
}
\arguments{
\item{data}{The data to be amputed}

\item{perc}{A scalar. The percentage (0-1) to be amputed.}

\item{cols}{The columns to ampute.}
}
\value{
The same dataset with random values set to NA in \code{vars}.
}
\description{
randomly imputes data.
}
\examples{
data(iris)
head(iris,10)

ampIris <- amputeData(iris)
head(ampIris,10)
}
