% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeData.R
\name{completeData}
\alias{completeData}
\title{completeData}
\usage{
completeData(miceObj, datasets = 1:miceObj$callParams$m, verbose = TRUE)
}
\arguments{
\item{miceObj}{an object of class miceDefs.}

\item{datasets}{a vector of the datasets you want to return.}

\item{verbose}{a warning is thrown if integers are converted to doubles.
To suppress this warning, set to \code{FALSE}.}
}
\value{
A list of imputed datasets.
}
\description{
Return imputed datasets from a \code{miceDefs} object.
}
\examples{
data("sampleMiceDefs")
imputedList <- completeData(sampleMiceDefs)
}
