% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarImps.R
\name{getVarImps}
\alias{getVarImps}
\title{getVarImps}
\usage{
getVarImps(miceObj, datasets = 1:miceObj$callParams$m, var)
}
\arguments{
\item{miceObj}{A miceDefs object created by \code{miceRanger}.}

\item{datasets}{The datasets to return. Can be a number, of a numeric vector.}

\item{var}{The variable to return the imputations for.}
}
\description{
Returns all of the imputations for the specified
datasets and variable.
}
\examples{
data("sampleMiceDefs")
getVarImps(sampleMiceDefs,"Sepal.Width")
}
