% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelError.R
\name{plotModelError}
\alias{plotModelError}
\title{plotModelError}
\usage{
plotModelError(
  miceObj,
  vars = names(miceObj$callParams$vars),
  pointSize = 1,
  ...
)
}
\arguments{
\item{miceObj}{an object of class miceDefs, created by the miceRanger function.}

\item{vars}{the variables you want to plot. Default is to plot all variables. Can be a vector of
variable names, or one of 'allNumeric' or 'allCategorical'}

\item{pointSize}{passed to \code{geom_point}, allows user to change dot size.}

\item{...}{other arguments passed to \code{grid.arrange()}}
}
\value{
nothing.
}
\description{
Plot the Out Of Bag model error for specified variables over all tierations.
}
\examples{
data("sampleMiceDefs")
plotModelError(sampleMiceDefs)
}
