% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarConvergence.R
\name{plotVarConvergence}
\alias{plotVarConvergence}
\title{plotVarConvergence}
\usage{
plotVarConvergence(miceObj, vars = miceObj$callParams$vars, ...)
}
\arguments{
\item{miceObj}{an object of class miceDefs, created by the miceRanger function.}

\item{vars}{the variables you want to plot. Default is to plot all variables. Can be a vector of
variable names, or one of 'allNumeric' or 'allCategorical'}

\item{...}{options passed to \code{grid.arrange()}}
}
\value{
nothing.
}
\description{
Create a impDefs object, which contains information about the imputation process.
}
\examples{
data(iris)
ampIris <- amputeData(iris)
miceObj <- miceRanger(
  ampIris
  , m = 2
  , maxiter = 3
  , verbose=FALSE
)
plotVarConvergence(miceObj)
}
