% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarImportance.R
\name{plotVarImportance}
\alias{plotVarImportance}
\title{plotVarImportance}
\usage{
plotVarImportance(
  miceObj,
  display = c("Relative", "Absolute"),
  dataset = 1,
  ...
)
}
\arguments{
\item{miceObj}{an object of class miceDefs, created by the miceRanger function.}

\item{display}{How do you want to display variable importance?
\itemize{
  \item {"Relative"} Scales the importance measure between 0-1 for each variable.
  \item {"Absolute"} Displays the variable importance as is. May be highly skewed.
}}

\item{dataset}{The dataset you want to plot the importance of.}

\item{...}{Other arguments passed to corrplot.}
}
\value{
nothing.
}
\description{
Create a impDefs object, which contains information about the imputation process.
}
\examples{
data("sampleMiceDefs")
plotVarImportance(sampleMiceDefs)
}
