% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.miceDefs}
\alias{print.miceDefs}
\title{Print a \code{miceDefs} object}
\usage{
\method{print}{miceDefs}(x, ...)
}
\arguments{
\item{x}{Object of class \code{miceDefs}}

\item{...}{required to use S3 method}
}
\value{
\code{NULL}
}
\description{
Print a \code{miceDefs} object
}
