% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDatasets.R
\name{addDatasets}
\alias{addDatasets}
\title{addDatasets}
\usage{
addDatasets(miceObj, datasets = 3, parallel = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{miceObj}{A miceDefs object created by \code{miceRanger}.}

\item{datasets}{The number of datasets to add.}

\item{parallel}{Should the process run in parallel? This process will take advantage of any cluster 
set up when \code{miceRanger} is called.}

\item{verbose}{should progress be printed?}

\item{...}{other parameters passed to \code{ranger()} to control model building.}
}
\value{
an updated miceDefs object with additional datasets.
}
\description{
Add datasets to a current miceDefs object. 
Adds the same number of iterations as other datasets.
}
\examples{
data("sampleMiceDefs")
miceObj <- addIterations(
    sampleMiceDefs
  , iters = 2
  , verbose = FALSE
  , num.threads = 1
  , num.trees=5
)
}
