% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeData.R
\name{completeData}
\alias{completeData}
\title{completeData}
\usage{
completeData(miceObj, datasets = 1:miceObj$callParams$m)
}
\arguments{
\item{miceObj}{an object of class miceDefs.}

\item{datasets}{a vector of the datasets you want to return.}
}
\value{
A list of imputed datasets.
}
\description{
Create a impDefs object, which contains information about the imputation process.
}
\examples{
data(iris)

ampIris <- amputeData(iris)

miceObj <- miceRanger(
  ampIris
  , m = 2
  , maxiter = 2
  , verbose=FALSE
)
imputedList <- completeData(miceObj)
}
