% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarImps.R
\name{getVarImps}
\alias{getVarImps}
\title{Get Variable Imputations}
\usage{
getVarImps(x, datasets, var)
}
\arguments{
\item{x}{A \code{miceDefs} or \code{impDefs} object.}

\item{datasets}{The datasets to return. Can be a number, of a numeric vector.}

\item{var}{The variable to return the imputations for.}
}
\value{
A matrix of imputations for a single variable. 
Each column represents a different dataset.
}
\description{
Returns imputations for the specified datasets and variable.
}
\details{
These functions exist solely to get at the imputed data for a
specific dataset and variable.
}
\examples{
data("sampleMiceDefs")
getVarImps(sampleMiceDefs,var="Sepal.Width")
}
