% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistributions.R
\name{plotDistributions}
\alias{plotDistributions}
\title{plotDistributions}
\usage{
plotDistributions(
  miceObj,
  vars = names(miceObj$callParams$vars),
  dotsize = 0.5,
  ...
)
}
\arguments{
\item{miceObj}{an object of class miceDefs, created by the miceRanger function.}

\item{vars}{the variables you want to plot. Default is to plot all variables. Can be a vector of
variable names, or one of 'allNumeric' or 'allCategorical'}

\item{dotsize}{Passed to \code{geom_dotplot()}. Depending on the number of graphs plotted, you may want to
change the dot size for categorical variables.}

\item{...}{additional parameters passed to \code{ggarrange()}.}
}
\value{
an object of class \code{ggarrange}.
}
\description{
Plots the distribution of the original data beside the imputed data.
}
\examples{
data("sampleMiceDefs")
plotDistributions(sampleMiceDefs)
}
