% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarImportance.R
\name{plotVarImportance}
\alias{plotVarImportance}
\title{plotVarImportance}
\usage{
plotVarImportance(
  miceObj,
  display = c("Relative", "Absolute"),
  dataset = 1,
  ...
)
}
\arguments{
\item{miceObj}{an object of class miceDefs, created by the miceRanger function.}

\item{display}{How do you want to display variable importance?
\itemize{
  \item {"Relative"} Scales the importance measure between 0-1 for each variable.
  \item {"Absolute"} Displays the variable importance as is. May be highly skewed.
}}

\item{dataset}{The dataset you want to plot the variable importance of.}

\item{...}{Other arguments passed to \code{corrplot()}.}
}
\description{
Plot the variable importance for each imputed variable. The values
represent the variables on the top axis importance in imputing the variables on
the left axis.
}
\examples{
data("sampleMiceDefs")
plotVarImportance(sampleMiceDefs)
}
