% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarImportance.R
\name{plotVarImportance}
\alias{plotVarImportance}
\title{plotVarImportance}
\usage{
plotVarImportance(
  miceObj,
  vars = miceObj$callParams$vars,
  title = NULL,
  display = c("Relative", "Absolute"),
  dataset = 1,
  ...
)
}
\arguments{
\item{miceObj}{an object of class miceDefs, created by the miceRanger function.}

\item{vars}{the variables you want to plot. Default is to plot all variables. Can be a vector of
variable names, or one of 'allNumeric' or 'allCategorical'}

\item{title}{The title of the plot. Default is no title.}

\item{display}{How do you want to display variable importance?
\itemize{
  \item {"Relative"} Scales the importance measure between 0-1 for each variable.
  \item {"Absolute"} Displays the variable importance as is. May be highly skewed.
}}

\item{dataset}{The dataset you want to plot the importance of.}

\item{...}{Other arguments passed to corrplot.}
}
\value{
nothing.
}
\description{
Create a impDefs object, which contains information about the imputation process.
}
\examples{
data(iris)
ampIris <- amputeData(iris)
miceObj <- miceRanger(
  ampIris
  , m = 2
  , maxiter = 3
  , verbose=FALSE
)
plotVarImportance(miceObj)
}
